// Atspm/WebUI/src/features/locations/types/Location.ts
// #region license
// Copyright 2026 Utah Department of Transportation
// for WebUI - Location.ts
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//http://www.apache.org/licenses/LICENSE-2.
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
// #endregion
import { Area } from '@/features/areas/types'
import { Device } from '@/features/devices/types'
import { Jurisdiction } from '@/features/jurisdictions/types'
import { Approach, LocationType } from '@/features/locations/types'
import { Region } from '@/features/regions/types'

export type Location = {
  areas: number[]
  latitude: number
  longitude: number
  primaryName: string
  locationIdentifier: string
  secondaryName: string
  jurisdictionId: number
  locationTypeId: number
  regionId: number
  chartEnabled: boolean
  charts: number[]
  start: string
  pedsAre1to1: boolean
  id: string
  note: string
  approaches?: Approach[]
  jurisdiction?: Jurisdiction
  region?: Region
}

export type LocationExpanded = Omit<Location, 'areas' | 'charts'> & {
  areas: Area[]
  approaches: Approach[]
  jurisdiction: Jurisdiction
  region: Region
  locationType: LocationType
  devices: Device[]
  note: string
  locationTypeId: number
}